module-3/assignment-3/pyproject.toml
[project]
name = "assignment-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.2.0",
    "python-dotenv>=1.0.1",
]
